langchain-framework/streamlitapp/requirements.txt
streamlit
langchain
langchain_chroma
langchain_community
python-dotenv
langchain_groq
langchain_huggingface
langchain_chroma
langchain_core
pypdf